% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{corp_name}
\alias{corp_name}
\title{Get Company Name}
\usage{
corp_name(x)
}
\arguments{
\item{x}{Glassdoor url}
}
\value{
Company name
}
\description{
Get Company Name
}
\examples{
\dontrun{
corp_name(glassdoor_url)
}
}
