% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dt_topic_table}
\alias{dt_topic_table}
\title{Datatable of Topic Terms}
\usage{
dt_topic_table(dat, company, direction)
}
\arguments{
\item{dat}{Dataset}

\item{company}{Company name}

\item{direction}{Pro or Con}
}
\value{
DT Datatable with Topics and top 12 terms
}
\description{
Datatable of Topic Terms
}
\examples{
\dontrun{
dt_topic_table(data, company, "pros")
}
}
