% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{estimate_max}
\alias{estimate_max}
\title{Estimate max}
\usage{
estimate_max(companyID)
}
\arguments{
\item{companyID}{Alpha numeric string from url}
}
\value{
Numeric value of total number of pages
}
\description{
Estimating the total number of pages to crawl
}
\examples{
\dontrun{
reviews <- estimate_max("E14069")
}
}
