% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{hc_sum_rating}
\alias{hc_sum_rating}
\title{Summary of Ratings}
\usage{
hc_sum_rating(dat)
}
\arguments{
\item{dat}{Glassdoor data}
}
\value{
Highcharter plot
}
\description{
Summary of Ratings
}
\examples{
\dontrun{
hc_sum_rating(data)
}
}
