% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{hc_topic_time}
\alias{hc_topic_time}
\title{Topic Prevalence over Time}
\usage{
hc_topic_time(dat)
}
\arguments{
\item{dat}{Glassdoor data}
}
\value{
Highcharter plot
}
\description{
Topic Prevalence over Time
}
\examples{
\dontrun{
hc_topic_time(data)
}
}
