% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{hc_wordcloud}
\alias{hc_wordcloud}
\title{Highcharter Wordcloud}
\usage{
hc_wordcloud(dat, type)
}
\arguments{
\item{dat}{Glassdoor data}

\item{type}{Summary, Pro or Con}
}
\value{
Highcharter wordcloud
}
\description{
Highcharter Wordcloud
}
\examples{
\dontrun{
hc_wordcloud(data, 'pros')
}
}
