% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{plot_hc_topics}
\alias{plot_hc_topics}
\title{Highcharter Plot}
\usage{
plot_hc_topics(dat, company, direction)
}
\arguments{
\item{dat}{Dataset}

\item{company}{Company name}

\item{direction}{Pros or Cons}
}
\value{
Highcharter plot
}
\description{
Highcharter Plot
}
\examples{
\dontrun{
plot_hc_topics(data, company_name, "pros")
}
}
