% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_glassdoor.R
\name{scrape_glassdoor}
\alias{scrape_glassdoor}
\title{Scrape Glassdoor Reviews}
\usage{
scrape_glassdoor(companyID, max)
}
\arguments{
\item{companyID}{String with alpha-numeric company ID}

\item{max}{Maximum number of pages to scrape}
}
\value{
Tibble review date, employee title, employee status, duration, summary, rating, pros, and cons
}
\description{
Function to scrape Glassdoor reviews and put into a tidy format
}
\examples{
\dontrun{
reviews <- scrape_glassdoor("E14069", max)
}
}
